moved:index and view to pages folder
